% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgui.R
\name{cgui}
\alias{cgui}
\title{cgui}
\usage{
cgui(envirfile = NULL)
}
\arguments{
\item{envirfile}{A file is given as input (type="file") or data object: possibly type={"list","table"}}

\item{type}{The object type of ff (type = {"file","list","table"})}
}
\description{
cgui is a GUI wrapper for converting betweem MPS data formats (sequencies) to LUS
}
\details{
The function produces a graphical window for letting the user define the settings
}
\author{
Oyvind Bleka
}
