% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(dat, lookupf = NULL, type = "LUS", toFile = FALSE,
  outf = "converted.csv", printMissing = TRUE,
  missf = "MissingSequences.csv", LUSsymb = "_")
}
\arguments{
\item{dat}{A data table with format [SampleName,Marker,Allele,Height] where Alleles are sequences. Height column is optional (indicating Evid if present and Ref if not present).}

\item{lookupf}{The destination/path of the UAS Lookup file (use build-in by default). Must be an xlsx file with same format. Not used if LUSstrR installed}

\item{type}{The convertion type (RU, LUS, LUS+)}

\item{toFile}{Boolean of whether the converted data table should be saved to workspace}

\item{outf}{The file name of the converted data table}

\item{printMissing}{A boolean of whether alleles not found in the look-up table should be printed to a seperate file}

\item{missf}{The file name of the missing alleles (not found in look-up table)}

\item{LUSsymb}{The separator format to use.}
}
\value{
Returning a table with format [SampleName,Marker,Allele,Height]
}
\description{
Sequence to LUS format converter
}
\details{
This script enables users to provide their own UAS sequence data and convert the sequences directly to different formats. The formats supported are RU (Repeat Unit/CE), LUS (Longest Uninterupted Stretch), LUS+ (extended version of LUS), or STRseq (The short format provided by STRseq)
The lookup table was created by Rebecca Just. The converter can also utilize LUSstrR if installed
}
\author{
Oyvind Bleka (in collaboration with Rebecca Just)
}
